move to createOrUpdateUser
